Rename all diff-diveinfo files to diff-diveresults
