+++ projects/planner/src/app/diff/diff.component.html
@@ -10,7 +10,7 @@
             <app-diff-waypoints></app-diff-waypoints>
         </div>
         <div class="col-xl-6 col-xxl-6 mt-4">
-            <app-diff-diveinfo></app-diff-diveinfo>
+            <app-diff-diveresults></app-diff-diveresults>
         </div>
     </div>
 </div>
