+++ projects/planner/src/app/diff/diveresults/diff-diveresults.component.ts
@@ -2,9 +2,9 @@ import {Component} from '@angular/core';
 import {faSlidersH} from '@fortawesome/free-solid-svg-icons';
 
 @Component({
-    selector: 'app-diff-diveinfo',
-    templateUrl: './diff-diveinfo.component.html',
-    styleUrls: ['./diff-diveinfo.component.scss']
+    selector: 'app-diff-diveresults',
+    templateUrl: './diff-diveresults.component.html',
+    styleUrls: ['./diff-diveresults.component.scss']
 })
 export class DiveInfoDifferenceComponent {
     public icon = faSlidersH;
