+++ projects/planner/src/app/diff/diveresults/results/diff-diveresults-results.component.scss
